refactor(core): stop routing Goal turns through the legacy Stop hook
Goal v3 (`goal-runtime`/`goal-reducer`/`goal-protocol`) owns every Goal
surface. The first-generation Stop-hook implementation is unreachable
from core: `registerGoalHook`'s only production caller is
`restoreGoalFromHistory`, which has no callers of its own.

Cut core's use of the legacy path so those modules become unreferenced:

- Re-home `interface ActiveGoal` from `goals/activeGoalStore.ts` into
  `goals/goal-legacy-projection.ts`, next to (but deliberately separate
  from) `LegacyActiveGoal` — the two differ in optionality, so merging
  them would change a wire type. `goals/index.ts` keeps exporting the
  name.
- `core/client.ts`: drop the permit-less `getActiveGoal` fallback and the
  `maybeEmitActiveGoalChange` duplicate-suppression helper it fed, the
  `abortGoalForStopHookCap` call on the blocking-cap path, and the
  `goalContinuationChanged`/`discardGoalContinuation` continuation
  rewrite. `projectActiveGoal`/`sameActiveGoalProjection` and the v3
  snapshot projection that backs the headless `active_goal` stream event
  are untouched.
- `config/config.ts`: drop the `GOAL_HOOK_ID_OUTPUT_KEY` special-casing
  of Stop hook outputs, along with the `hasNonGoalBlockingStopHook` and
  `nonGoalBlockingStopReason` response fields that only the legacy
  continuation rewrite read.

The legacy files themselves stay for a follow-up.
