+++ packages/core/src/core/client.test.ts
@@ -101,12 +101,6 @@ import {
 import { collectAvailableSkillEntries } from '../tools/skill-utils.js';
 import type { AvailableSkillEntry } from '../tools/skill-utils.js';
 import { ToolNames } from '../tools/tool-names.js';
-import {
-  __resetActiveGoalStoreForTests,
-  clearActiveGoal,
-  setActiveGoal,
-} from '../goals/activeGoalStore.js';
-import { GOAL_HOOK_ID_OUTPUT_KEY } from '../goals/goalHook.js';
 import { emptyGoalSnapshot } from '../goals/goal-protocol.js';
 import type { GoalRuntime } from '../goals/goal-runtime.js';
 import type { FileHistorySnapshot } from '../services/fileHistoryService.js';
@@ -728,7 +722,6 @@ describe('Gemini Client (client.ts)', () => {
   afterEach(() => {
     vi.useRealTimers();
     vi.restoreAllMocks();
-    __resetActiveGoalStoreForTests();
   });
 
   describe('initialize', () => {
@@ -11634,186 +11627,19 @@ Other open files:
         client['chat'] = mockChat as LlmChat;
       });
 
-      it('emits active_goal when a goal is active for the turn', async () => {
-        setActiveGoal('test-session-id', {
-          condition: 'finish the refactor',
-          iterations: 2,
-          setAt: 123,
-          tokensAtStart: 456,
-          hookId: 'goal-hook-id',
-          lastReason: 'still missing verification',
-        });
-
-        const events = await fromAsync(
-          client.sendMessageStream(
-            [{ text: 'Hi' }],
-            new AbortController().signal,
-            'prompt-active-goal',
-          ),
-        );
-
-        expect(events[0]).toEqual({
-          type: LlmEventType.ActiveGoal,
-          value: {
-            condition: 'finish the refactor',
-            iterations: 2,
-            setAt: 123,
-            tokensAtStart: 456,
-            hookId: 'goal-hook-id',
-            lastReason: 'still missing verification',
-          },
-        });
-      });
-
-      it('emits active_goal null when the Stop hook clears the goal', async () => {
-        setActiveGoal('test-session-id', {
-          condition: 'finish the refactor',
-          iterations: 2,
-          setAt: 123,
-          tokensAtStart: 456,
-          hookId: 'goal-hook-id',
-          lastReason: 'still missing verification',
-        });
-        const mockMessageBus = {
-          request: vi.fn().mockImplementation(async () => {
-            clearActiveGoal('test-session-id');
-            return {};
-          }),
-          response: vi.fn(),
-        };
-        vi.mocked(mockConfig.getDisableAllHooks).mockReturnValue(false);
-        vi.mocked(mockConfig.getMessageBus).mockReturnValue(
-          mockMessageBus as unknown as ReturnType<Config['getMessageBus']>,
-        );
-        vi.mocked(mockConfig.hasHooksForEvent).mockImplementation(
-          (event: string) => event === 'Stop',
-        );
-        client['chat'] = {
-          addHistory: vi.fn(),
-          getHistory: vi.fn().mockReturnValue([
-            {
-              role: 'model',
-              parts: [{ text: 'done' }],
-            },
-          ]),
-        } as unknown as LlmChat;
-        mockTurnRunFn.mockReturnValue(
-          (async function* () {
-            yield { type: LlmEventType.Content, value: 'done' };
-          })(),
-        );
-
-        const events = await fromAsync(
-          client.sendMessageStream(
-            [{ text: 'Hi' }],
-            new AbortController().signal,
-            'prompt-cleared-active-goal',
-          ),
-        );
-
-        expect(events).toContainEqual({
-          type: LlmEventType.ActiveGoal,
-          value: null,
-        });
-      });
-
-      it('emits active_goal null when the Stop hook clears the goal before aborting', async () => {
-        setActiveGoal('test-session-id', {
-          condition: 'finish the refactor',
-          iterations: 2,
-          setAt: 123,
-          tokensAtStart: 456,
-          hookId: 'goal-hook-id',
-          lastReason: 'still missing verification',
-        });
+      it('does not start a Stop hook continuation when the blocking decision lands already aborted', async () => {
         const abortController = new AbortController();
         const mockMessageBus = {
-          request: vi.fn().mockImplementation(async () => {
-            clearActiveGoal('test-session-id');
-            abortController.abort();
-            return {};
-          }),
-          response: vi.fn(),
-        };
-        vi.mocked(mockConfig.getDisableAllHooks).mockReturnValue(false);
-        vi.mocked(mockConfig.getMessageBus).mockReturnValue(
-          mockMessageBus as unknown as ReturnType<Config['getMessageBus']>,
-        );
-        vi.mocked(mockConfig.hasHooksForEvent).mockImplementation(
-          (event: string) => event === 'Stop',
-        );
-        client['chat'] = {
-          addHistory: vi.fn(),
-          getHistory: vi.fn().mockReturnValue([
-            {
-              role: 'model',
-              parts: [{ text: 'done' }],
-            },
-          ]),
-        } as unknown as LlmChat;
-        mockTurnRunFn.mockReturnValue(
-          (async function* () {
-            yield { type: LlmEventType.Content, value: 'done' };
-          })(),
-        );
-
-        const events = await fromAsync(
-          client.sendMessageStream(
-            [{ text: 'Hi' }],
-            abortController.signal,
-            'prompt-cleared-active-goal-then-aborted',
-          ),
-        );
-
-        const activeGoalEvents = events.filter(
-          (event) => event.type === LlmEventType.ActiveGoal,
-        );
-
-        expect(activeGoalEvents).toEqual([
-          {
-            type: LlmEventType.ActiveGoal,
-            value: expect.objectContaining({
-              condition: 'finish the refactor',
-            }),
-          },
-          {
-            type: LlmEventType.ActiveGoal,
-            value: null,
-          },
-        ]);
-      });
-
-      it('emits active_goal changes when aborting before Stop hook continuation', async () => {
-        setActiveGoal('test-session-id', {
-          condition: 'finish the refactor',
-          iterations: 2,
-          setAt: 123,
-          tokensAtStart: 456,
-          hookId: 'goal-hook-id',
-          lastReason: 'still missing verification',
-        });
-        const abortController = new AbortController();
-        const mockMessageBus = {
-          request: vi.fn().mockImplementation(async () => {
-            setActiveGoal('test-session-id', {
-              condition: 'finish the refactor',
-              iterations: 3,
-              setAt: 123,
-              tokensAtStart: 456,
-              hookId: 'goal-hook-id',
-              lastReason: 'still missing validation',
-            });
-            return {
-              output: {
-                get decision() {
-                  abortController.abort();
-                  return 'block';
-                },
-                reason: 'Keep working',
+          request: vi.fn().mockImplementation(async () => ({
+            output: {
+              get decision() {
+                abortController.abort();
+                return 'block';
               },
-              stopHookCount: 1,
-            };
-          }),
+              reason: 'Keep working',
+            },
+            stopHookCount: 1,
+          })),
           response: vi.fn(),
         };
         vi.mocked(mockConfig.getDisableAllHooks).mockReturnValue(false);
@@ -11845,27 +11671,10 @@ Other open files:
             'prompt-stop-hook-continuation-aborted',
           ),
         );
-        const activeGoalEvents = events.filter(
-          (event) => event.type === LlmEventType.ActiveGoal,
-        );
 
-        expect(activeGoalEvents).toEqual([
-          {
-            type: LlmEventType.ActiveGoal,
-            value: expect.objectContaining({
-              condition: 'finish the refactor',
-              iterations: 2,
-            }),
-          },
-          {
-            type: LlmEventType.ActiveGoal,
-            value: expect.objectContaining({
-              condition: 'finish the refactor',
-              iterations: 3,
-              lastReason: 'still missing validation',
-            }),
-          },
-        ]);
+        // The blocking decision arrives only after the signal aborted, so the
+        // continuation turn must not run and no StopHookLoop is announced.
+        expect(mockTurnRunFn).toHaveBeenCalledOnce();
         expect(events).not.toContainEqual(
           expect.objectContaining({
             type: LlmEventType.StopHookLoop,
@@ -12503,74 +12312,6 @@ Other open files:
         ).toHaveBeenCalledWith('ok', { promptId: 'prompt-stop-hook-budget' });
       });
 
-      it('emits one active_goal null when the blocking cap aborts an active goal', async () => {
-        setActiveGoal('test-session-id', {
-          condition: 'finish the refactor',
-          iterations: 2,
-          setAt: 123,
-          tokensAtStart: 456,
-          hookId: 'goal-hook-id',
-          lastReason: 'still missing verification',
-        });
-        const mockMessageBus = {
-          request: vi.fn().mockResolvedValue({
-            output: {
-              decision: 'block',
-              reason: 'Keep working',
-            },
-            stopHookCount: 1,
-          }),
-          response: vi.fn(),
-        };
-        vi.mocked(mockConfig.getDisableAllHooks).mockReturnValue(false);
-        vi.mocked(mockConfig.getMessageBus).mockReturnValue(
-          mockMessageBus as unknown as ReturnType<Config['getMessageBus']>,
-        );
-        vi.mocked(mockConfig.hasHooksForEvent).mockImplementation(
-          (event: string) => event === 'Stop',
-        );
-        vi.mocked(mockConfig.getStopHookBlockingCap).mockReturnValue(1);
-
-        client['chat'] = {
-          addHistory: vi.fn(),
-          getHistory: vi.fn().mockReturnValue([
-            {
-              role: 'model',
-              parts: [{ text: 'not done' }],
-            },
-          ]),
-        } as unknown as LlmChat;
-        mockTurnRunFn.mockReturnValue(
-          (async function* () {
-            yield { type: LlmEventType.Content, value: 'not done' };
-          })(),
-        );
-
-        const events = await fromAsync(
-          client.sendMessageStream(
-            [{ text: 'Hi' }],
-            new AbortController().signal,
-            'prompt-stop-cap-active-goal',
-          ),
-        );
-        const activeGoalEvents = events.filter(
-          (event) => event.type === LlmEventType.ActiveGoal,
-        );
-
-        expect(activeGoalEvents).toEqual([
-          {
-            type: LlmEventType.ActiveGoal,
-            value: expect.objectContaining({
-              condition: 'finish the refactor',
-            }),
-          },
-          {
-            type: LlmEventType.ActiveGoal,
-            value: null,
-          },
-        ]);
-      });
-
       it('should not skip hooks when hasHooksForEvent returns true', async () => {
         const mockMessageBus = {
           request: vi.fn().mockResolvedValue({ modifiedPrompt: undefined }),
@@ -13079,259 +12820,6 @@ Other open files:
         expect(mockTurnRunFn).toHaveBeenCalledTimes(2);
         expect(getLastTurnRequestText()).toContain('Keep working');
         expect(getLastTurnRequestText()).toContain('also check the tests');
-      });
-
-      it('preserves goal feedback alongside an external stop reason', async () => {
-        setActiveGoal('test-session-id', {
-          condition: 'finish the refactor',
-          iterations: 1,
-          setAt: 123,
-          tokensAtStart: 456,
-          hookId: 'goal-hook',
-        });
-        const mockMessageBus = {
-          request: vi
-            .fn()
-            .mockResolvedValueOnce({
-              output: {
-                decision: 'block',
-                continue: false,
-                stopReason: 'External stop hook feedback',
-                reason: 'Keep working on the active goal',
-                hookSpecificOutput: {
-                  [GOAL_HOOK_ID_OUTPUT_KEY]: 'goal-hook',
-                },
-              },
-              stopHookCount: 2,
-              hasNonGoalBlockingStopHook: true,
-              nonGoalBlockingStopReason: 'External stop hook feedback',
-            })
-            .mockResolvedValue({ output: undefined }),
-          response: vi.fn(),
-        };
-        vi.mocked(mockConfig.getDisableAllHooks).mockReturnValue(false);
-        vi.mocked(mockConfig.getMessageBus).mockReturnValue(
-          mockMessageBus as unknown as ReturnType<Config['getMessageBus']>,
-        );
-        vi.mocked(mockConfig.hasHooksForEvent).mockImplementation(
-          (event: string) => event === 'Stop',
-        );
-        mockTurnRunFn.mockImplementation(() =>
-          (async function* () {
-            yield { type: LlmEventType.Content, value: 'response' };
-          })(),
-        );
-        const getSteerInput = vi
-          .fn<() => Promise<SteerInput | undefined>>()
-          .mockResolvedValue(undefined);
-
-        await fromAsync(
-          client.sendMessageStream(
-            [{ text: 'start the goal' }],
-            new AbortController().signal,
-            'prompt-goal-with-external-stop-reason',
-            { type: SendMessageType.UserQuery, getSteerInput },
-          ),
-        );
-
-        expect(mockTurnRunFn).toHaveBeenCalledTimes(2);
-        expect(getLastTurnRequestText()).toContain(
-          'External stop hook feedback',
-        );
-        expect(getLastTurnRequestText()).toContain(
-          'Keep working on the active goal',
-        );
-      });
-
-      it('stops a blocking goal when queued input clears it', async () => {
-        setActiveGoal('test-session-id', {
-          condition: 'finish the refactor',
-          iterations: 1,
-          setAt: 123,
-          tokensAtStart: 456,
-          hookId: 'old-goal-hook',
-        });
-        const mockMessageBus = {
-          request: vi.fn().mockResolvedValue({
-            output: {
-              decision: 'block',
-              reason: 'Keep working',
-              hookSpecificOutput: {
-                [GOAL_HOOK_ID_OUTPUT_KEY]: 'old-goal-hook',
-              },
-            },
-            stopHookCount: 2,
-            hasNonGoalBlockingStopHook: false,
-          }),
-          response: vi.fn(),
-        };
-        vi.mocked(mockConfig.getDisableAllHooks).mockReturnValue(false);
-        vi.mocked(mockConfig.getMessageBus).mockReturnValue(
-          mockMessageBus as unknown as ReturnType<Config['getMessageBus']>,
-        );
-        vi.mocked(mockConfig.hasHooksForEvent).mockImplementation(
-          (event: string) => event === 'Stop',
-        );
-        mockTurnRunFn.mockImplementation(() =>
-          (async function* () {
-            yield { type: LlmEventType.Content, value: 'response' };
-          })(),
-        );
-        const getSteerInput = vi
-          .fn<() => Promise<SteerInput | undefined>>()
-          .mockResolvedValueOnce(undefined)
-          .mockImplementationOnce(async () => {
-            clearActiveGoal('test-session-id');
-            return undefined;
-          });
-
-        const events = await fromAsync(
-          client.sendMessageStream(
-            [{ text: 'start the goal' }],
-            new AbortController().signal,
-            'prompt-clear-during-stop',
-            { type: SendMessageType.UserQuery, getSteerInput },
-          ),
-        );
-
-        expect(mockTurnRunFn).toHaveBeenCalledOnce();
-        expect(events).toContainEqual({
-          type: LlmEventType.ActiveGoal,
-          value: null,
-        });
-      });
-
-      it('replaces a blocking goal without sending the old continuation', async () => {
-        setActiveGoal('test-session-id', {
-          condition: 'finish the refactor',
-          iterations: 1,
-          setAt: 123,
-          tokensAtStart: 456,
-          hookId: 'old-goal-hook',
-        });
-        const mockMessageBus = {
-          request: vi
-            .fn()
-            .mockResolvedValueOnce({
-              output: {
-                decision: 'block',
-                reason: 'Keep working',
-                hookSpecificOutput: {
-                  [GOAL_HOOK_ID_OUTPUT_KEY]: 'old-goal-hook',
-                },
-              },
-              stopHookCount: 1,
-              hasNonGoalBlockingStopHook: false,
-            })
-            .mockResolvedValue({ output: undefined }),
-          response: vi.fn(),
-        };
-        vi.mocked(mockConfig.getDisableAllHooks).mockReturnValue(false);
-        vi.mocked(mockConfig.getMessageBus).mockReturnValue(
-          mockMessageBus as unknown as ReturnType<Config['getMessageBus']>,
-        );
-        vi.mocked(mockConfig.hasHooksForEvent).mockImplementation(
-          (event: string) => event === 'Stop',
-        );
-        mockTurnRunFn.mockImplementation(() =>
-          (async function* () {
-            yield { type: LlmEventType.Content, value: 'response' };
-          })(),
-        );
-        const getSteerInput = vi
-          .fn<() => Promise<SteerInput | undefined>>()
-          .mockResolvedValueOnce(undefined)
-          .mockImplementationOnce(async () => {
-            setActiveGoal('test-session-id', {
-              condition: 'verify the tests',
-              iterations: 0,
-              setAt: 789,
-              tokensAtStart: 999,
-              hookId: 'new-goal-hook',
-            });
-            return {
-              parts: [{ text: 'new goal instruction' }],
-              accept: vi.fn(),
-              restore: vi.fn(),
-            };
-          })
-          .mockResolvedValue(undefined);
-
-        await fromAsync(
-          client.sendMessageStream(
-            [{ text: 'start the goal' }],
-            new AbortController().signal,
-            'prompt-replace-during-stop',
-            { type: SendMessageType.UserQuery, getSteerInput },
-          ),
-        );
-
-        expect(mockTurnRunFn).toHaveBeenCalledTimes(2);
-        expect(getLastTurnRequestText()).toContain('new goal instruction');
-        expect(getLastTurnRequestText()).not.toContain('Keep working');
-      });
-
-      it('preserves other blocking Stop hook output when a goal is cleared', async () => {
-        setActiveGoal('test-session-id', {
-          condition: 'finish the refactor',
-          iterations: 1,
-          setAt: 123,
-          tokensAtStart: 456,
-          hookId: 'old-goal-hook',
-        });
-        const mockMessageBus = {
-          request: vi
-            .fn()
-            .mockResolvedValueOnce({
-              output: {
-                decision: 'block',
-                reason: 'Keep working\nPolicy review is still required',
-                hookSpecificOutput: {
-                  [GOAL_HOOK_ID_OUTPUT_KEY]: 'old-goal-hook',
-                },
-              },
-              stopHookCount: 2,
-              hasNonGoalBlockingStopHook: true,
-              nonGoalBlockingStopReason: 'Policy review is still required',
-            })
-            .mockResolvedValue({ output: undefined }),
-          response: vi.fn(),
-        };
-        vi.mocked(mockConfig.getDisableAllHooks).mockReturnValue(false);
-        vi.mocked(mockConfig.getMessageBus).mockReturnValue(
-          mockMessageBus as unknown as ReturnType<Config['getMessageBus']>,
-        );
-        vi.mocked(mockConfig.hasHooksForEvent).mockImplementation(
-          (event: string) => event === 'Stop',
-        );
-        mockTurnRunFn.mockImplementation(() =>
-          (async function* () {
-            yield { type: LlmEventType.Content, value: 'response' };
-          })(),
-        );
-        const getSteerInput = vi
-          .fn<() => Promise<SteerInput | undefined>>()
-          .mockResolvedValueOnce(undefined)
-          .mockImplementationOnce(async () => {
-            clearActiveGoal('test-session-id');
-            return undefined;
-          })
-          .mockResolvedValue(undefined);
-
-        await fromAsync(
-          client.sendMessageStream(
-            [{ text: 'start the goal' }],
-            new AbortController().signal,
-            'prompt-clear-with-other-stop-hook',
-            { type: SendMessageType.UserQuery, getSteerInput },
-          ),
-        );
-
-        expect(mockTurnRunFn).toHaveBeenCalledTimes(2);
-        expect(getLastTurnRequestText()).toContain(
-          'Policy review is still required',
-        );
-        expect(getLastTurnRequestText()).not.toContain('Keep working');
       });
 
       it('uses input queued during next-speaker classification for the continuation', async () => {
