+++ packages/core/src/goals/index.ts
@@ -5,7 +5,6 @@
  */
 
 export type {
-  ActiveGoal,
   GoalTerminalEvent,
   GoalTerminalKind,
   GoalTerminalObserver,
@@ -57,6 +56,7 @@ export {
   projectGoalStateToLegacy,
 } from './goal-legacy-projection.js';
 export type {
+  ActiveGoal,
   LegacyActiveGoal,
   LegacyGoalProjection,
   LegacyGoalStatus,
